requirements.txt
playwright
bs4
requests
pandas
streamlit
geopy
scikit-learn
folium
streamlit-folium
Pillow

